SCHEMA-749. Fix memory leak in layout watching.
When we create a layout tracker, there are some series of references where one of them is a watcher that gets passed to zookeeper (a long lived thread). This watcher has a reference (since it's a non-static inner) to its parent object which has a non-null reference to the layout tracker which is effectively the HBaseKijiTable. In a multiple classloader environment, this prevents the parent classloader from being unloaded and hence a memory leak.

By nulling out the layout tracker reference on close, we break this dependency only leaving a "lightweight" reference to the watcher. Until we can remove watchers, this will continue to be the case.

Review: https://review.kiji.org/r/1789/
Reviewer: milo